refactor(associations): relocate through-association target loads to their Rails homes
