+++ packages/activerecord/src/associations/polymorphic-sti-through.test.ts
@@ -7,8 +7,8 @@
  *   - `has_many :through` whose source reflection is a polymorphic
  *     belongs_to, disambiguated by `source_type:` ("polymorphic
  *     has_many through"). The fixture layers this on top of a
- *     nested through (Hotel → Departments → Chefs), so `findTarget`
- *     routes through `loadHasManyThrough`'s walker rather than the
+ *     nested through (Hotel → Departments → Chefs), so `HasManyAssociation#findTarget`
+ *     routes through `HasManyThroughAssociation#findTarget`'s walker rather than the
  *     final-step JOIN/AssociationScope path. Both that walker and
  *     the `includes()` preloader must filter through-records by the
  *     polymorphic discriminator (`*_type`) and only materialize the
