+++ packages/activerecord/src/index.ts
@@ -50,7 +50,6 @@ export {
   registerModel,
   modelRegistry,
   loadHasOne,
-  loadHasManyThrough,
   association,
   isAssociationCached,
   updateCounterCaches,
